Add purge-old-kernels state
The script was lifted from the bikeshed package to be able to
install this without needing the 125MB+ of other bikeshed stuff.

Might be useful for environments where the kernel is upgradable, like
your own hardware. Cloud environemnts mostly lock down the kernel, so
less of a problem with /boot filling up there. Also, at least on DO,
/boot is not a separate partition, thus unlikely to fill up easily.