Run postgres.server state before most other states
Needed to ensure the database is present. Other states shouldn't depend
on it explicitly, both since the actual backend might be flexible, but
also since most cases the server would be on a different host, thus the
service would be unavailable to wait for. Thus this is primarily a fix
for vagrant and other self-contained systems running the database on the
same host as the service.