powerdns: Remove unused exit through floating ip
First of all, this wasn't used, secondly the digitialocean grains
caused root to try http, external grains and/or pillar values should
rather be put behind cachish or similar to make it easier to segment
this data. Lastly, this would only work if we knew all the replica IPs
through the salt system, but since replicas can be configured through
the powerdns database this assumption doesn't hold.