Enable setting proxy_pass path
Without this it's not possible to trim the prefix when forwarding.
Usually a location block like

  location /prefix {
    proxy_pass https://example.com;
  }

Causes `GET /prefix` to end up as `GET /prefix` also for the backend.
However, if the proxy_pass statement ends with a path, the prefix is
trimmed from the forwarding request. Thus a block like

  location /prefix {
    proxy_pass https://example.com/;
  }

(note the trailing slash) will make a request like `GET /prefix`
translate to `GET /` on the backend. This change preserves the path set
in the upstream in the proxy_pass statement, which enables controlling
this behavior.